update handler of downloadCliAction
