+++ Utils/azure-toolkit-ide-libs/azure-toolkit-ide-common-lib/src/main/resources/bundles/com/microsoft/azure/toolkit/operation.properties
@@ -328,6 +328,7 @@ docker.push_image=push docker image to web app within run/debug configuration
 docker.start_app=start docker based web app within run/debug configuration
 docker.deploy_image=deploy docker image to web app
 function.run_app.module=run function app module ({0}) locally
+function.run_app=run function app locally
 connector.setup_connection_for_configuration.config=setup resource connection(s) for run configuration ({0})
 webapp.create_app_properties_editor.app=create properties editor for Azure Web App ({0})
 function.deploy_app.module=deploy module ({0}) function app
